requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
PyGithub==2.4.0
httpx==0.27.2
google-generativeai==0.8.3
pydantic==2.9.2
pandas==2.2.0
openpyxl==3.1.2
